Add trigger scintillator pad sim hits DQM analyzer and sample config
Currently, no trigger pad digi exists in master, so this commit only
adds a DQM analyzer for sim hits. (Mostly, there is nothing implemented
for PEs or anything to do with noise.)

There is a hacky, temporary fix to the EcalHexReadout: default gap 0-->1
in anticipation of a real fix to this (making it configurable).

A sample config file, to be run with ldmx-app and no arguments, is added.
It currently assumes simulation has already been run. Tweak input file name
accordingly. At this stage, I can't get both of these steps running in the
same sequence. I might get to debugging that, but it's not central to having
trigger pad DQM available for validation.